allure_test_template/allure_test_template/requirements.txt
airtest==1.2.7
airtest-selenium==1.0.3
allure-pytest==2.10.0
allure-python-commons==2.10.0
Appium-Python-Client==2.4.0
async-generator==1.10
atomicwrites==1.4.1
attrs==22.1.0
cached-property==1.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
chromedriver-py==105.0.5195.19
colorama==0.4.5
comtypes==1.1.14
cryptography==38.0.1
Cython==0.29.28
decorator==5.1.1
deprecate-kwargs==0.0.2
Deprecated==1.2.13
docx==0.2.4
et-xmlfile==1.1.0
ez-setup==0.9
facebook-wda==1.4.6
gensim==4.2.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.2
logging42==0.0.8
loguru==0.6.0
lxml==4.9.1
MarkupSafe==2.1.1
MouseInfo==0.1.3
mss==6.1.0
numpy==1.23.2
opencv-contrib-python==4.6.0.66
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.5.2
pdfminer==20191125
Pillow==9.2.0
pip==22.2.2
pluggy==1.0.0
py==1.11.0
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodome==3.15.0
PyGetWindow==0.0.9
PyJWT==2.5.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
pynput==1.7.6
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.20.0
pytweening==1.0.4
pytz==2022.6
pywin32==304
pywinauto==0.6.3
PyYAML==6.0
requests==2.28.1
retry==0.9.2
rin-exceptions==1.0.3
scipy==1.9.1
selenium==4.4.3
six==1.16.0
smart-open==6.1.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.3
win32-setctime==1.1.0
wrapt==1.14.1
wsproto==1.2.0